Fixed freeze from compiling man-o-war blueprint while actor was in current level in editor
For those of you having issues compiling the man-o-war blueprint:

When you open a level in the editor, it creates a game world for the spawned actors that you see in the view port. If you modify one of those actor's blueprint significantly, it DESTROYS the actor, and recreates it. This normally would not be an issue except for the thumbnail system:

void FClassActorThumbnailScene::ClearStaleActors()

This loops through the CURRENT actors and destroys them.

HOWEVER when you destroy the man-o-war, "Event Destroyed" fires in blueprints....
And inside of that we:
SPAWN ANOTHER man-o-war

adding it to the list that ClearStaleActors() is referencing. This causes an infinite loop of new actors being created then destroyed ect ect.

Simply remove the spawn node from the "Event Destroyed"!!!!!!!!!!!!!!!!! (This needs to actually be moved to gamemode or player controller)